Fixes for https://sourceforge.net/p/widgets-95/tickets/1/ ... fixes GCC build for recent, untested, last-minute commits. Includes some new changes also.

+++ src/glui_edittext.cpp
@@ -282,7 +282,7 @@ void Text_Interface::set_limits(double l1, double l2, int limit_type)
 
 	if(data_type!=UI_EDIT_TEXT)
 	{
-		float new_val = float_val; CLAMP(new_val,l1,l2);
+		double new_val = float_val; CLAMP(new_val,l1,l2);
 
 		if(new_val!=float_val) set_float_val(new_val);
 	}
